Assign hostname to ActionPlan when picked-up
Currently, the hostname for an ActionPlan is saved when the execution is
started and moved to ONGOING.

That is actually missleading as the host is actually assigned to the
ActionPlan when the RPC is picked by the applier and the execution is
added to the execution pool.

While in many cases the time to start the execution since picked-up is
almost instantaneous, there are cases where this time can be longer, as for
example when the ThreadPoolExecutor is full and the applier needs to
wait for an existing execution before starting the next one. In those
cases we need to identify the Applier service which owns the Pending
ActionPlan so it is more correct to set the hostname in the ActionPlan
as soon as it is assigned to the applier.

Implements: blueprint-monitor-failed-appliers
Change-Id: Icd08fd1b7626adc0acb2948def7bc5b3f8eb9a8a
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>