Retry calls to nova API when connection errors are detected
Currently, when a call to the Nova API fails for any connection
temporary issue, Watcher actions fail.

This patch is adding retries to nova api calls when connection issues
are detected. The retries can be configured with new parameters in the
nova section:

- http_retries (default is 3 retries)
- http_retry_interval (default is 2 seconds).

Note that, in order to implement the retries in all the nova_helper
methods, in some cases i am adding new elementary calls which are called
by more complex ones.

Closes-Bug: #2133934

Change-Id: I587920f9e287cec2df3489b13c4ef78013de1876
Signed-off-by: Alfredo Moralejo <amoralej@redhat.com>